fix: api를 query 파일로 합치기
